world/tick.go: Fixed ticking breaking when setting negative numbers.

+++ server/world/tick.go
@@ -150,7 +150,7 @@ func (t ticker) tickBlocksRandomly(loaders []*Loader, tick int64) {
 
 		// We generate up to j random positions for every sub chunk.
 		x, y, z := g.uint4(t.w.r), g.uint4(t.w.r), g.uint4(t.w.r)
-		for j := uint32(0); j < uint32(t.w.conf.RandomTickSpeed); j++ {
+		for j := 0; j < t.w.conf.RandomTickSpeed; j++ {
 			for i, sub := range c.Sub() {
 				if sub.Empty() {
 					// SubChunk is empty, so skip it right away.
